websec_tool/requirements.txt
requests
selenium
python-owasp-zap-v2.4
beautifulsoup4
